[refacotr] 불필요한 QuizDataText 파일 제거(#196)
