--- feature/quiz/src/main/java/kr/boostcamp_2024/course/quiz/component/QuizDataText.kt
@@ -1,34 +1,0 @@
-package kr.boostcamp_2024.course.quiz.component
-
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextOverflow
-import kr.boostcamp_2024.course.domain.model.BaseQuiz
-
-@Composable
-fun QuizDataText(
-    quiz: BaseQuiz?,
-) {
-    quiz?.let { quiz ->
-        // QuizTitle
-        Text(
-            text = quiz.title,
-            style = MaterialTheme.typography.displayMedium,
-            fontWeight = FontWeight.SemiBold,
-            color = Color.White,
-            maxLines = 1,
-            overflow = TextOverflow.Ellipsis,
-        )
-
-        // QuizDescription
-        quiz.description?.let { description ->
-            Text(
-                text = description,
-                color = Color.White,
-            )
-        }
-    }
-}
