+++ feature/quiz/src/main/java/kr/boostcamp_2024/course/quiz/presentation/quiz/QuizScreen.kt
@@ -10,9 +10,11 @@ import androidx.compose.foundation.layout.calculateStartPadding
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
@@ -22,6 +24,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalLayoutDirection
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
@@ -35,7 +39,6 @@ import kr.boostcamp_2024.course.domain.model.Quiz
 import kr.boostcamp_2024.course.domain.model.RealTimeQuiz
 import kr.boostcamp_2024.course.quiz.component.QuizDataButton
 import kr.boostcamp_2024.course.quiz.component.QuizDataChip
-import kr.boostcamp_2024.course.quiz.component.QuizDataText
 import kr.boostcamp_2024.course.quiz.component.QuizTopAppBar
 import kr.boostcamp_2024.course.quiz.viewmodel.QuizViewModel
 
@@ -182,6 +185,31 @@ fun QuizScreen(
                     onWaitingRealTimeQuizButtonClick = onWaitingRealTimeQuizButtonClick,
                 )
             }
+        }
+    }
+}
+
+@Composable
+fun QuizDataText(
+    quiz: BaseQuiz?,
+) {
+    quiz?.let { quiz ->
+        // QuizTitle
+        Text(
+            text = quiz.title,
+            style = MaterialTheme.typography.displayMedium,
+            fontWeight = FontWeight.SemiBold,
+            color = Color.White,
+            maxLines = 1,
+            overflow = TextOverflow.Ellipsis,
+        )
+
+        // QuizDescription
+        quiz.description?.let { description ->
+            Text(
+                text = description,
+                color = Color.White,
+            )
         }
     }
 }
